requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
jira>=3.5.0
gspread>=5.10.0
google-auth>=2.0.0
